# ENH: special: Add root finding tools to xsf and implement gdtrib for GPU while making substantial improvements