--- a/scipy/special/cython_special.pyx
+++ b/scipy/special/cython_special.pyx
@@ -1123,7 +1123,7 @@ cdef extern from r"xsf_wrappers.h":
     double special_beip(double) nogil
     double special_ber(double) nogil
     double special_berp(double) nogil
-    double special_gdtrib(double, double, double) nogil
+    double xsf_gdtrib(double, double, double) nogil
     npy_double special_kei(npy_double) nogil
     npy_double special_keip(npy_double) nogil
     void special_ckelvin(npy_double, npy_cdouble *, npy_cdouble *, npy_cdouble *, npy_cdouble *) nogil
@@ -2650,7 +2650,7 @@ cpdef double gdtria(double x0, double x1, double x2) noexcept nogil:
 
 cpdef double gdtrib(double x0, double x1, double x2) noexcept nogil:
     """See the documentation for scipy.special.gdtrib"""
-    return special_gdtrib(x0, x1, x2)
+    return xsf_gdtrib(x0, x1, x2)
 
 cpdef double gdtrix(double x0, double x1, double x2) noexcept nogil:
     """See the documentation for scipy.special.gdtrix"""