--- a/scipy/special/meson.build
+++ b/scipy/special/meson.build
@@ -302,7 +302,8 @@ xsf_sources = [
   'xsf/wright_bessel.h',
   'xsf/hyp2f1.h',
   'xsf/tools.h',
-  'xsf/iv_ratio.h'
+  'xsf/iv_ratio.h',
+  'xsf/cdflib.h'
 ]
 
 xsf_cephes_sources = [
@@ -340,7 +341,8 @@ xsf_cephes_sources = [
   'xsf/cephes/ellie.h',
   'xsf/cephes/ellik.h',
   'xsf/cephes/ellpe.h',
-  'xsf/cephes/ellpk.h'
+  'xsf/cephes/ellpk.h',
+  'xsf/cephes/igam_asymp_coeff.h'
 ]
 
 py3.install_sources(xsf_sources, subdir: 'scipy/special/xsf')