--- a/scipy/special/xsf/cdflib.h
+++ b/scipy/special/xsf/cdflib.h
@@ -59,8 +59,8 @@ XSF_HOST_DEVICE inline double gdtrib(double a, double p, double x) {
     }
     double lower_bound = std::numeric_limits<double>::min();
     double upper_bound = std::numeric_limits<double>::max();
-    auto [x_left, x_right, bracket_status] =
-        detail::bracket_root(func, 1.0, 5.0, lower_bound, upper_bound, 8.0, false, 1000, args);
+    auto [xl, xr, f_xl, f_xr, bracket_status] =
+        detail::bracket_root_for_cdf_inversion(func, 0.1, 1.0, lower_bound, upper_bound, 10.0, false, 100, args);
     if (bracket_status == 1) {
         set_error("gdtrib", SF_ERROR_UNDERFLOW, NULL);
         return 0.0;
@@ -74,7 +74,9 @@ XSF_HOST_DEVICE inline double gdtrib(double a, double p, double x) {
         ;
         return std::numeric_limits<double>::quiet_NaN();
     }
-    auto [result, root_status] = detail::find_root_bus_dekker_r(func, x_left, x_right, 1000, args);
+    auto [result, root_status] = detail::find_root_chandrupatla(
+        func, xl, xr, f_xl, f_xr, std::numeric_limits<double>::epsilon(), 1e-100, 100, args
+    );
     if (root_status) {
         /* The root finding return should only fail if there's a bug in our code. */
         set_error("gdtrib", SF_ERROR_OTHER, "Computational Error, (%.17g, %.17g, %.17g)", a, p, x);