--- a/scipy/special/xsf/cdflib.h
+++ b/scipy/special/xsf/cdflib.h
@@ -41,27 +41,17 @@ XSF_HOST_DEVICE inline double gdtrib(double a, double p, double x) {
         /* gdtrib(a, p, x) tends to 0 as p -> 1.0 from the left when x > 0. */
         return 0.0;
     }
-
     double q = 1.0 - p;
-    double (*func)(double, std::tuple<double, double, double>);
-    std::tuple<double, double, double> args;
-    if (p <= q) {
-        args = {a, p, x};
-        func = +[](double b, std::tuple<double, double, double> args) -> double {
-            auto [a, p, x] = args;
-            return cephes::igam(b, a * x) - p;
-        };
-    } else {
-        args = {a, q, x};
-        func = +[](double b, std::tuple<double, double, double> args) -> double {
-            auto [a, q, x] = args;
-            return q - cephes::igamc(b, a * x);
-        };
-    }
+    auto func = [a, p, q, x](double b) -> double {
+	if (p <= q) {
+	    return cephes::igam(b, a * x) - p;
+	}
+	return q - cephes::igamc(b, a * x);
+    };
     double lower_bound = std::numeric_limits<double>::min();
     double upper_bound = std::numeric_limits<double>::max();
     auto [xl, xr, f_xl, f_xr, bracket_status] = detail::bracket_root_for_cdf_inversion(
-        func, 1.0, lower_bound, upper_bound, -0.875, 7.0, 0.125, 8, false, 342, args
+        func, 1.0, lower_bound, upper_bound, -0.875, 7.0, 0.125, 8, false, 342
     );
     if (bracket_status == 1) {
         set_error("gdtrib", SF_ERROR_UNDERFLOW, NULL);
@@ -73,16 +63,14 @@ XSF_HOST_DEVICE inline double gdtrib(double a, double p, double x) {
     }
     if (bracket_status >= 3) {
         set_error("gdtrib", SF_ERROR_OTHER, "Computational Error");
-        ;
         return std::numeric_limits<double>::quiet_NaN();
     }
     auto [result, root_status] = detail::find_root_chandrupatla(
-        func, xl, xr, f_xl, f_xr, std::numeric_limits<double>::epsilon(), 1e-100, 100, args
+        func, xl, xr, f_xl, f_xr, std::numeric_limits<double>::epsilon(), 1e-100, 100
     );
     if (root_status) {
         /* The root finding return should only fail if there's a bug in our code. */
         set_error("gdtrib", SF_ERROR_OTHER, "Computational Error, (%.17g, %.17g, %.17g)", a, p, x);
-        ;
         return std::numeric_limits<double>::quiet_NaN();
     }
     return result;