… scipy/special/functions.json
… scipy/special/tests/test_xsf_cuda.py
… scipy/special/xsf/config.h
… scipy/special/xsf_wrappers.cpp
… scipy/special/xsf_wrappers.h