<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message.
However, please only include an issue number in the description, not the title,
and please ensure that any code names containing underscores are enclosed in backticks.

Depending on your changes, you can skip CI operations and save time and energy: 
http://scipy.github.io/devdocs/dev/contributor/continuous_integration.html#skipping

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### Reference issue
<!--Example: Closes gh-WXYZ.-->
This will tick a box for https://github.com/scipy/scipy/issues/20223

#### What does this implement/fix?
<!--Please explain your changes.-->

##### Root finding

This PR implements scalar root-finding tools in xsf that we can use to bring functions from cdflib from the GPU. Rather than trying to translate the root finding tools from the current cdflib C translation, I rewrote them from scratch, following the Algol code in the paper [Two Efficient Algorithms with Guaranteed Convergence for Finding a Zero of a Function](https://dl.acm.org/doi/10.1145/355656.355659) by Bus and Dekker, for the implementation of the root finding algorithm used in cdflib, and writing an initial bracket finder based on the general idea from @mdhaber's [bracket_root](https://github.com/scipy/scipy/blob/4d2c9e050663efd54396a3d8749792fad0b55dde/scipy/optimize/_elementwise.py#L462-L624), which itself uses well understood bracket finding techniques. I made some adjustments to accommodate searching over the entire range of double values. Both the root finding and bracket finding routines assume monotonicity.

The implementation of root finding tools in cdflib is hopelessly tangled, due its use of a reverse communication style which is a workaround for Fortran 77's lack of callbacks. I think the new implementation is much easier to follow, and because I'm able to understand, I was able to make some improvements. I had pointed out [here](https://github.com/scipy/scipy/pull/19824#issuecomment-1902098317) that the cdflib implementation doesn't do a good job of letting you control convergence. The rate of convergence outspeeds the reported tolerance, so tolerance was set at 1e-8 originally, because setting the tolerance too low often caused additional jumps that lead to a worse solution. The new implementation uses a criterion which allows us to search to the limits of machine precision, which will be seen in the improvement in the quality of results.

##### gdtrib

I have implemented `gdtrib` from cdflib as a proof of concept. The work in this PR is only a small fraction of the work I've done relating to cdflib. I've translated much of it to C++, and benchmarked it's routines against those from cephes. Rather than using the incomplete gamma function routine from cdflib, I've opted to use `igam` and `igamc` from cephes, which are comparable in accuracy, and have the benefit of not allocating arrays internally; which means they can be used on the GPU. Cephes is also what is used in `gdtr`, so by using cephes, `gdtrib` becomes more of a proper inverse with respect to `b` of our `gdtr`. I think it would be good to follow a general principle that our inverse cdf functions invert the actual implementations we use for the `cdf`.

##### Unable to use std::function

One thing that is annoying, is that I couldn't have functions be passed to the root finding routines as `std::function`, and use capturing lambda's to construct functions within the inverse cdf functions to pass to the root finders. In order to make this work in CuPy using NVRTC, I needed to use function pointers of type `double (*func)(double, std::tuple<Args...>)`, where the tuple of args is for the distribution parameters that we'd like to be able to just capture in a lambda. The root finding routine itself takes a tuple of args, `std::tuple<Args...> args`, which we can use to specialize the function. I hope we can eventually get `std::function` and capturing lambda's to work in CuPy, but I think using function pointers like is a reasonable workaround for now.

##### Exceptional cases
I mentioned in https://github.com/scipy/scipy/issues/20972#issuecomment-2267261963, that I'd like to try to fix the edge cases as I translate `cdflib`. In this regard, the changes I've made here are

* Expand the search space for the parameter `b` to all normalized positive doubles. Previously the search space was the interval `[1e-100, 1e100]`. The bracket finding algorithm I've written has no issues with the expanded search space.
* If the bracket finding algorithm determines that the actual root must be greater than the largest double, infinity is returned, and overflow is signaled with `set_error`. If it determines that the actual root must be smaller than the smallest normalized double, underflow is signaled and zero is returned.
* Limiting cases are handled explicitly. These are
    *  `gdtrib(a, p, x)` tends to 0 as `x -> 0` when `p > 0`.
    * `gdtrib(a, p, x)` tends to infinity as `p -> 0` from the right when `x > 0`.
    * `gdtrib(a, p, x)` tends to 0 as `p -> 1.0` from the left when x > 0`.
    * `gdtrib(a, 0, 0)` is indeterminate`
    I had noticed for instance that for the `p = 0` and `p = 1` cases, there will be a range of `b` values where `gdtr(a, b, x)` numerically evaluates to 0 or 1 respectively, and the old implementation would just return one of them, almost at random, depending upon the vagaries of how it's initial bracket finder is set up.

##### CuPy
I've verified that this works in CuPy and have added a test case to `scipy/special/tests/test_xsf_cuda.py`.

#### Additional information
<!--Any additional information you think is important.-->
##### GPU Speedup
I used the following script to create a CuPy `ElementwiseKernel` for `gdtrib` and set up some large arrays to run it over, to compare timing with `gdtrib` on the CPU

```python
import cupy
import numpy as np
import scipy.special as sc



preamble = "#include <xsf/cdflib.h>"
routine = "out0 = xsf::gdtrib(in0, in1, in2)"
name = "gdtrib"
in_params = "float64 in0, float64 in1, float64 in2"
out_params = "float64 out0"


cupy_gdtrib = cupy.ElementwiseKernel(
    in_params, out_params, routine, name, preamble=preamble,
    options=(f"--include-path={sc._get_include()}", "-std=c++17"),
)

a = cupy.array(2.5)
p = cupy.linspace(1e-10, 1, 3000)
x = cupy.linspace(1e-10, 10.0, 3000)
p, x = cupy.meshgrid(p, x)

A = 2.5
P = np.linspace(1e-10, 1, 3000)
X = np.linspace(1e-10, 10, 3000)
P, X = np.meshgrid(P, X)

# Calling the function will invoke the JIT to compile the ElementwiseKernel. We don't compilation to be
# included in the timing
cupy_gdtrib(a, p[0], x[0])
```

and observed a roughly 5x speedup in Ipython using my NVIDIA RTX A4000 GPU.

```
In [2] %time sc.gdtrib(A, P, X)
CPU times: user 31.2 s, sys: 0 ns, total: 31.2 s
Wall time: 31.2 s
```

```
In [3] %time cupy_gdtrib(a, p, x).get()
CPU times: user 6.29 s, sys: 11.7 ms, total: 6.3 s
Wall time: 6.3 s
```

##### Improvements in accuracy

I wrote a very rough script which I used to compare the new and old implementations along the following metrics

1. Accuracy of `gdtr(a, gdtrib(a, p, x), x)` in estimating `p`.
2. Accuracy of `mp_gdtr(a, gdtrib(a, p, x), x)` in estimating `p`. Where `mp_gdtr` is a correctly rounded version of `gdtr` which uses arbitrary precision internally and converts to float before returning.
3. Comparison of `gdtrib(a, p, x)` against an arbitrary precision reference.

Currently the old `gdtrib` remains in this branch under the name `gdtrib_old`, solely for benchmarking purposes. We'll need to take this out before this get merged.

<details>

First the script I used to generate the data

```python
import numpy as np
import pandas as pd

from mpmath import mp
from scipy.special import gdtrib, gdtrib_old, gdtr


def mp_gdtrib(a, p, x, guess):
    a, p, x = mp.mpf(a), mp.mpf(p), mp.mpf(x)
    def func(b):
        result = mp.gammainc(b, 0, a*x, regularized=True) - p
        return result
    with mp.workdps(100):
        try:
            result = float(
                mp.findroot(
                    func,
                    guess,
                    verify=True,
                    tol=1e-30,
                )
            )
        except Exception:
            result = np.nan
    return result


def round_trip(p, a, x):
    b = gdtrib(a, p, x)
    return gdtr(a, b, x)

def round_trip_old(p, a, x):
    b = gdtrib_old(a, p, x)
    return gdtr(a, b, x)

def mp_round_trip(p, a, x):
    b = gdtrib(a, p, x)
    with mp.workdps(30):
        try:
            result = float(mp.gammainc(b, 0, a*x, regularized=True))
        except Exception:
            return np.nan
    return result


def mp_round_trip_old(p, a, x):
    b = gdtrib_old(a, p, x)
    with mp.workdps(30):
        try:
            result = float(mp.gammainc(b, 0, a*x, regularized=True))
        except Exception:
            return np.nan
    return result


def func(a, p, x, b):
    q = 1 - p
    if p <= q:
        return gammainc(b, a*x) - p
    return q - gammaincc(b, a*x)


def neighbors(x, k):
    output = []
    horizon = np.inf
    if k < 0:
        horizon = -horizon
        k = -k
    for i in range(k):
        x = np.nextafter(x, horizon)
        output.append(x)
    return np.asarray(output)


rng = np.random.default_rng(1729)

P = np.hstack(
    (
        rng.uniform(0, 1, size=10),
        neighbors(1.0, -3), # very close to 1.0,
    )
)
P = np.sort(P)
A = rng.uniform(0.8, 1.2, size=16) * 2.0**np.arange(-4, 12)
X = np.hstack(
    (rng.uniform(0.8, 1.2, size=16) * 2.0**np.arange(-4, 12))
)

# gdtrib
p_array, a_array, x_array = np.meshgrid(P, A, X)
p_array, a_array, x_array = p_array.ravel(), a_array.ravel(), x_array.ravel()

results1 = []
results2 = []
expected = []
p_rt, p_rt_old, p_rt_mp, p_rt_mp_old = [], [], [], []
for p, a, x in zip(p_array, a_array, x_array):
    val1 = gdtrib(a, p, x)
    val2 = gdtrib_old(a, p, x)
    expected.append(mp_gdtrib(a, p, x, val2))
    results1.append(val1)
    results2.append(val2)
    p_rt.append(round_trip(p, a, x))
    p_rt_old.append(round_trip_old(p, a, x))
    p_rt_mp.append(mp_round_trip(p, a, x))
    p_rt_mp_old.append(mp_round_trip_old(p, a, x))

results1 = np.asarray(results1)
results2 = np.asarray(results2)
expected = np.asarray(expected)
p_rt = np.asarray(p_rt)
p_rt_old = np.asarray(p_rt_old)
p_rt_mp = np.asarray(p_rt_mp)
p_rt_mp_old = np.asarray(p_rt_old)

df = pd.DataFrame(
    {
        "p": p_array,
        "a": a_array,
        "x": x_array,
        "results1": results1,
        "results2": results2,
        "expected": expected,
        "p_rt": p_rt,
        "p_rt_old": p_rt_old,
        "p_rt_mp": p_rt_mp,
        "p_rt_mp_old": p_rt_mp_old,
    }
)
```

and another for generating the tables

```python
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd


df = pd.read_csv("gdtrib_results.csv", sep=",")


columns = ["branch", "min"] + [f"{n}%" for n in range(10, 91, 10)] + ["max"]
rt_re_old = abs(df.p_rt_old - df.p) / abs(df.p)
rt_re = abs(df.p_rt - df.p) / abs(df.p)
rt_qt_old = np.quantile(rt_re_old, np.linspace(0, 1, 11))
rt_qt = np.quantile(rt_re, np.linspace(0, 1, 11))


table1 = pd.DataFrame(
    [
        ["main"] + rt_qt_old.tolist(),
        ["PR"] + rt_qt.tolist(),
    ],
    columns=columns,
).to_markdown(index=False, tablefmt="gh")

cc = df.dropna()

rt_re_mp = abs(cc.p_rt_mp - cc.p) / abs(cc.p)
rt_re_mp_old = abs(cc.p_rt_mp_old - cc.p) / abs(cc.p)
rt_mp_qt = np.quantile(rt_re_mp, np.linspace(0, 1, 11))
rt_mp_qt_old = np.quantile(rt_re_mp_old, np.linspace(0, 1, 11))

table2 = pd.DataFrame(
    [
        ["main"] + rt_mp_qt_old.tolist(),
        ["PR"] + rt_mp_qt.tolist(),
    ],
    columns=columns,
).to_markdown(index=False, tablefmt="gh")
                    

re_mp = abs(cc.results1 - cc.expected) / abs(cc.expected)
re_mp_old = abs(cc.results2 - cc.expected) / abs(cc.expected)
mp_qt = np.quantile(re_mp, np.linspace(0, 1, 11))
mp_qt_old = np.quantile(re_mp_old, np.linspace(0, 1, 11))

table3 = pd.DataFrame(
    [
        ["main"] + mp_qt_old.tolist(),
        ["PR"] + mp_qt.tolist(),
    ],
    columns=columns,
).to_markdown(index=False, tablefmt="gh")
```

</details>


This first table shows percentiles for relative error when comparing `gdtr(a, gdtrib(a, p, x), x)` against `p` over 3328 test cases from the script above.

| branch   |   min |   10% |   20% |         30% |         40% |         50% |         60% |         70% |         80% |         90% |         max |
|----------|-------|-------|-------|-------------|-------------|-------------|-------------|-------------|-------------|-------------|-------------|
| main     |     0 |     0 |     0 | 4.94289e-16 | 2.46548e-15 | 1.00135e-14 | 9.07437e-14 | 9.52067e-13 | 8.6112e-12  | 8.03292e-11 | 5.75493e-08 |
| PR       |     0 |     0 |     0 | 0           | 0           | 2.27422e-16 | 4.09993e-16 | 7.70651e-16 | 2.15851e-15 | 6.20768e-15 | 4.21913e-13 |

The `mpmath` arbitrary precision reference for `gdtr` wasn't able to converge for some of the test cases. The following tables are for a subset of 3276 of the above test cases.

Here is the round trip comparison where `gdtr` is replaced with a correctly rounded reference implementation. The results are comparable. This isn't surprising because `gdtr` already evaluates at close to machine precision.

| branch   |   min |   10% |   20% |         30% |         40% |         50% |         60% |         70% |         80% |         90% |         max |
|----------|-------|-------|-------|-------------|-------------|-------------|-------------|-------------|-------------|-------------|-------------|
| main     |     0 |     0 |     0 | 5.50885e-16 | 2.50164e-15 | 1.00989e-14 | 8.84778e-14 | 9.0074e-13  | 8.20264e-12 | 7.57065e-11 | 4.06397e-08 |
| PR       |     0 |     0 |     0 | 0           | 0           | 1.36664e-16 | 3.77691e-16 | 6.59052e-16 | 1.69597e-15 | 5.27242e-15 | 2.11494e-13 |

Finally here is the comparison to the arbitrary precision reference

| branch   |   min |   10% |         20% |         30% |         40% |         50% |         60% |         70% |         80% |         90% |         max |
|----------|-------|-------|-------------|-------------|-------------|-------------|-------------|-------------|-------------|-------------|-------------|
| main     |     0 |     0 | 2.12641e-16 | 6.11437e-16 | 1.61093e-15 | 6.23451e-15 | 3.77914e-14 | 2.63097e-13 | 1.86824e-12 | 9.09311e-12 | 4.91479e-11 |
| PR       |     0 |     0 | 0           | 0           | 0           | 0           | 1.31742e-16 | 1.89296e-16 | 2.64941e-16 | 5.95813e-16 | 1.41677e-13 |

The improvement is dramatic.

##### Next steps
Although only one function from cdflib is made available on the GPU here, most of the work towards making this possible is in this PR, and after this gets in, it will be tedious but straightforward to follow through and make the rest of cdflib work on the GPU. 

Also, unrelated to this PR, I've found that we can significantly improve our implementation of the `gamma` function by combining the `cephes` and `cdflib` implementations, as suggested by @ilayn here https://github.com/scipy/scipy/pull/19824#issue-2069115062. I will eventually get around to this in a followup.

cc @rlucas7 for review, @izaid in case you want to take a look at this, @fancidev for any suggestions you might have, @WarrenWeckesser since you've been active in `special` lately and might be interested in this.