@ilayn 2024-08-26
Did not really sit down and lay out the differences but if this is working, it is a major improvement. Congrats @steppi ! Much appreciated. 

> One thing that is annoying, is that I couldn't have functions be ...

I know you don't like these comments but this is precisely how things go with C++, slowly things creep into the codebase because some very obvious thing is missing or impossible to do and we end up with more C++ 😅 They should have called it `std::mold` because you can't help but import more janky stuff `std::` to do very straightforward things.