@steppi 2024-08-26
> I know you don't like these comments but this is precisely how things go with C++, slowly things creep into the codebase because some very obvious thing is missing or impossible to do and we end up with more C++ 😅 They should have called it `std::mold` because you can't help but import more janky stuff `std::` to do very straightforward things.

On the contrary. The way I ended up passing functions here is pretty much idiomatic C (except I use non-capturing lamda’s to define the functions to find roots of instead of defining them outside of the cdf functions). This is a matter of us not getting to use a nice C++ feature due to the limitations of NVRTC. I think @izaid knows how to get `std::function` and capturing lamdas to work on GPU with NVRTC though. It’s more of a matter of my inexperience with GPU programming I think.