@fancidev 2024-08-27
> Also, I will say that handling “plateaus” like I did is a performance optimization for cases where the actual root is very far from the boundaries of the initial bracket. And yes, plateau should be set back to `false`.

Thanks for explaining. Indeed for use cases where the root is very far away from the initial bracket, the “plateau” handling is a viable optimization.

I tried to get a better understanding of the plateau handling behavior. Let $c$ denote `factor`. Let $(a_0,b_0)$ be first the bracket where $f(a_0)=f(b_0)$, so plateau handling kicks in. Let $\Delta=b_0-a_0$. I find that the following brackets, $(a_k,b_k)$, $k=1,2,3,\cdots$ have $a_k=a_0$, $b_k =a_0+(1+c)^k\Delta$, until some $b_k$ where $f(b_k)\ne f(a_0)$. This is where the function “goes out” of the plateau. If $f(b_k)$ has the opposite sign of $f(a_0)$, then it looks like $(a_0,b_k)$ is returned, which is probably a large interval.

In comparison, if plateau does not kick in, then I find $a_k=b_{k-1}$, $b_k=a_0+[(c^{k+1}-1)/(c-1)]\Delta$.

So the bracket size increment rate difference between non-plateau and plateau is basically a factor of $c$ versus a factor of $(1+c)$. Therefore I think the code could be simplified by multiplying the step size by `(1+factor)` if `y_frontier == y_interior` or by `factor` otherwise. There would be no need to “fix” `interior` in the plateau case, and the `plateau` variable could be dropped.