@ev-br 2024-08-27
Am wondering if 

- is this useful as a general-purpose root-finder?
- are you planning to make this routine public at some point?
- can one make definite statements comparing this routine with `brentq`, `fsolve` and other root-finders in scipy.optimize?