@steppi 2024-08-27
@fancidev, sorry, what I wrote about the plateau handling above is nonsense. I think there was a genuine pathological case that was fixed by not moving the trailing endpoint, but I can't remember what it was now, and won't have time to check until tomorrow morning UTC+8.