@fancidev 2024-08-30
> Thanks @fancidev for the thorough review. I think we have this in good shape now. Let me know if you have any more suggestions.

Glad to help! Looks perfect to me now!