@fancidev 2024-09-05
I guess the choice between `std::function` versus typed argument is primarily a trade-off between runtime speed and binary size. I don’t have any judgement about such trade-off, but indeed providing a typed argument makes it possible for the caller to pass a `std::function` if they want but not the other way around.