@rlucas7 2024-09-06
@steppi I may not have time to review this code as much (as I would like to) before Sept 20th-a little over two weeks from now. If it can wait, I'd love to dig into the details here, if now maybe someone else has time beforehand to keep this effort moving forward. I'm glad to take a look that weekend but I've got some commitments before then I need to do.