@mdhaber 2024-09-08
Just thought I'd collect a few `scipy.special` issues that could be addressed by root finding:
- https://github.com/scipy/scipy/issues/17507
- https://github.com/scipy/scipy/issues/15633
- https://github.com/scipy/scipy/issues/7695

Those might be good ones to take this infrastructure for a test spin (after it merges).