@steppi 2024-09-09
Thanks for the review @izaid! Great suggestions. I was able to banish the function pointers by putting the `if/else` inside of `func` instead of outside of it. I've verified locally that `gdtrib` is working on GPU and there was no reduction in performance on CPU or GPU.