@mdhaber 2024-09-10
Although there are many ways to test, of course I had to compare against our `optimize.elementwise.bracket_root`/`optimize.elementwise.find_root`  : ) With default settings on those, `gdtrib` in this PR is approximately twice as fast (on CPU) and has the same worst-case relative error.

```python3
import numpy as np
from scipy import special
from scipy.optimize import elementwise

a = 2.5
p = np.linspace(1e-10, 1, 300)
x = np.linspace(1e-10, 10, 300)
p, x = np.meshgrid(p, x)

def f(b, a, p, x):
    return special.gdtr(a, b, x) - p

args = (a, p, x)
bracket = elementwise.bracket_root(f, xl0=1, xmin=0, args=args)
root = elementwise.find_root(f, bracket.bracket, args=args)

# %timeit bracket = elementwise.bracket_root(f, xl0=1, xmin=0, args=args); root = elementwise.find_root(f, bracket.bracket, args=args)
# 435 ms ± 3.73 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# %timeit special.gdtrib(a, p, x)
# 222 ms ± 824 μs per loop (mean ± std. dev. of 7 runs, 1 loop each)

np.max(f(root.x, a, p, x)/p)  #  np.float64(5.0535565548161535e-14)
np.max(f(special.gdtrib(a, p, x), a, p, x)/p)  # np.float64(4.239300639334267e-14)
```